Renamed botocore_connector helper import to aiobotocore_connector
